+++ client/js/client.js
@@ -1,4 +1,0 @@
-const socket = io('http://localhost:3000');
-socket.on('connect', (data) => {
-    console.log(data);
-})