+++ CHANGELOG.md
@@ -9,5 +9,6 @@ All notable changes to the "Chat Program" project will be documented in this fil
 
 - Initial project structure.
 - Initial socket, express, ejs as template and insert a chat application ejs temp
+- add emit server message to clients and some changes
 
 ---